Add tests for entire step against reference
