+++ romancal/refpix/tests/test_data.py
@@ -1,12 +1,5 @@
-from enum import IntEnum
-
 import numpy as np
-import pytest
-from astropy import units as u
 from numpy.testing import assert_allclose
-from roman_datamodels.datamodels import RampModel, RefpixRefModel
-from roman_datamodels.maker_utils import mk_ramp
-from roman_datamodels.testing.factories import create_refpix_ref
 
 from romancal.refpix.data import (
     ChannelView,
@@ -17,85 +10,7 @@ from romancal.refpix.data import (
 )
 
 from . import reference_utils
-
-RNG = np.random.default_rng(42)
-
-
-class Dims(IntEnum):
-    """
-    Assumed data dimensions for dat (aid clarity of testing).
-        - N_FRAMES: the number of frames used in the test data (arbitrary) choice
-        - N_ROWS: the number of rows of the test data
-        - N_COLS: the number of columns of the test data
-    """
-
-    N_FRAMES = 8
-    N_ROWS = Const.N_COLUMNS
-    N_COLS = Const.N_COLUMNS + Const.CHAN_WIDTH
-    N_CHAN = N_COLS // Const.CHAN_WIDTH
-    PADDED_WIDTH = Const.CHAN_WIDTH + Const.PAD
-
-    N_FFT_COLS = (N_ROWS * PADDED_WIDTH) // 2 + 1
-
-
-@pytest.fixture(scope="module")
-def data():
-    return RNG.uniform(1, 100, size=(Dims.N_FRAMES, Dims.N_ROWS, Dims.N_COLS)).astype(
-        np.float32
-    )
-
-
-@pytest.fixture(scope="module")
-def datamodel(data):
-    datamodel = mk_ramp()
-
-    detector = data[:, :, : Const.N_COLUMNS].copy()
-    amp33 = data[:, :, Const.N_COLUMNS :].copy()
-
-    # Sanity check, that this is in line with the datamodel maker_utils
-    assert detector.shape == (Dims.N_FRAMES, Dims.N_ROWS, Const.N_COLUMNS)
-    assert amp33.shape == (Dims.N_FRAMES, Dims.N_ROWS, Const.CHAN_WIDTH)
-
-    datamodel.data = u.Quantity(
-        detector, unit=datamodel.data.unit, dtype=datamodel.data.dtype
-    )
-    datamodel.amp33 = u.Quantity(
-        amp33, unit=datamodel.amp33.unit, dtype=datamodel.amp33.dtype
-    )
-
-    datamodel.border_ref_pix_left = datamodel.data[:, :, : Const.REF]
-    datamodel.border_ref_pix_right = datamodel.data[:, :, -Const.REF :]
-
-    return RampModel(datamodel)
-
-
-@pytest.fixture(scope="module")
-def standard(data):
-    return StandardView(data)
-
-
-@pytest.fixture(scope="module")
-def channels(standard):
-    return standard.channels
-
-
-@pytest.fixture(scope="module")
-def coeffs() -> Coefficients:
-    coeffs = RNG.uniform(1, 100, size=(6, Const.N_DETECT_CHAN, Dims.N_FFT_COLS))
-
-    coeffs = coeffs[:3, :, :] + 1.0j * coeffs[3:, :, :]
-
-    return Coefficients(coeffs[0], coeffs[1], coeffs[2])
-
-
-@pytest.fixture(scope="module")
-def offset() -> np.ndarray:
-    return RNG.uniform(0, 100, size=(Dims.N_ROWS, Dims.N_COLS)).astype(np.float32)
-
-
-@pytest.fixture(scope="module")
-def ref_pix_ref():
-    return RefpixRefModel(create_refpix_ref())
+from .conftest import RNG, Dims
 
 
 def test_constants_sanity():
